Fluid Sprite Movement (rudimentary)
Except for a small stutter as the viewport snaps to tile boundaries,
this works fairly well.  It would benefit from a map+screen combination
refactor, since map and screen know WAY too much about each other, but
it (kinda) works.

It would also be nice to have an actual clock governing the sprite
animations, and some metadata embedded in the sprite files themselves.